refactor(api): Moved task helpers to api/api per https://github.com/armory-io/terraform-provider-spinnaker/pull/1#discussion_r239679819
