+++ spinnaker/api/application.go
@@ -103,33 +103,3 @@ func DeleteApplication(client *gate.GatewayClient, applicationName string) error
 
 	return nil
 }
-
-func taskCompleted(task map[string]interface{}) bool {
-	taskStatus, exists := task["status"]
-	if !exists {
-		return false
-	}
-
-	COMPLETED := [...]string{"SUCCEEDED", "STOPPED", "SKIPPED", "TERMINAL", "FAILED_CONTINUE"}
-	for _, status := range COMPLETED {
-		if taskStatus == status {
-			return true
-		}
-	}
-	return false
-}
-
-func taskSucceeded(task map[string]interface{}) bool {
-	taskStatus, exists := task["status"]
-	if !exists {
-		return false
-	}
-
-	SUCCESSFUL := [...]string{"SUCCEEDED", "STOPPED", "SKIPPED"}
-	for _, status := range SUCCESSFUL {
-		if taskStatus == status {
-			return true
-		}
-	}
-	return false
-}
